requirements.txt
# [Configuration Management]
pydantic==2.12.0
pydantic-settings==2.11.0
pyyaml==6.0.3

# [Logging]
colorlog==6.7.0

# [RAG and Embeddings]
sentence-transformers>=2.7.0
chromadb==0.5.5
boto3==1.34.0
requests==2.31.0
ollama==0.6.0